chore(authz-gamma-ui): remove dead code (TargetPickerDialog, AiModelTargetPickerDialog, GaplPreview, Tabs)
Three target-picker dialogs collapsed to one in UnifiedTargetPickerDialog
some time back, but the two superseded ones (TargetPickerDialog and
AiModelTargetPickerDialog) were never deleted. GaplPreview was never
wired into PolicyEditorSheet — the editor uses its own inline CodePane.
components/Tabs was scaffold leftover with no consumers in the module.

Verified no production references:
  - TargetPickerDialog → only self-test referenced it
  - AiModelTargetPickerDialog → only its AiProviderEntry type was still
    referenced from ServicePolicyPage and service-defs/llms.ts. Moved the
    type to service-defs/llms.ts (its natural home).
  - GaplPreview → only self-test
  - components/Tabs → zero importers

~774 LOC dead code gone. Build (`nx build gravitee-gamma-module-authz`)
still passes.
